refactor: manage user-code release through helm
