+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "dagster-uc"
-version = "0.3.5"
+version = "0.4.0"
 authors = [
     {name = "Stefan Verbruggen"},
     {name = "Ion Koutsouris"},
